# docs/lyc-api-openapi.yaml
openapi: 3.0.3
info:
  title: LYC 小宇宙学堂 API
  description: LYC 儿童启蒙学习平台后端 API，涵盖认证、内容管理、媒体播放、AI 分析、学习进度等模块。
  version: 1.0.0
  contact:
    name: LYC Team
servers:
  - url: http://127.0.0.1:3002
    description: 本地开发
  - url: https://api.lyc.example.com
    description: 生产环境
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 小程序端登录后获取的 JWT Token
    adminKey:
      type: apiKey
      in: header
      name: X-Admin-Key
      description: 管理员 API Key
  schemas:
    User:
      type: object
      properties:
        id:
          type: string
          example: "a1b2c3d4"
        openid:
          type: string
          nullable: true
        phoneNumber:
          type: string
          nullable: true
          example: "138****0001"
        nickname:
          type: string
          example: "微信用户"
        avatarUrl:
          type: string
          nullable: true
        gender:
          type: string
          enum: [male, female, undisclosed]
        role:
          type: string
          enum: [parent]
        phoneVerified:
          type: boolean
        profileCompleted:
          type: boolean
        status:
          type: string
          enum: [active, disabled]
        lastLoginAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
    ChildProfile:
      type: object
      properties:
        id:
          type: string
        nickname:
          type: string
          example: "小星星"
        birthYear:
          type: integer
          nullable: true
          example: 2020
        birthMonth:
          type: integer
          nullable: true
          example: 6
        gender:
          type: string
          enum: [male, female, undisclosed]
        avatarUrl:
          type: string
          nullable: true
        isPrimary:
          type: boolean
...
    Error:
      type: object
      properties:
        error:
          type: string
paths:
...
  /api/v1/auth/wx-login:
    post:
      tags: [认证]
      summary: 微信登录
      description: 使用微信 wx.login 获取的 code 进行登录，自动注册新用户。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [code]
              properties:
                code:
                  type: string
                  description: wx.login 返回的 code
      responses:
        "200":
          description: 登录成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                  user:
                    $ref: "#/components/schemas/User"
                  children:
                    type: array
                    items:
                      $ref: "#/components/schemas/ChildProfile"
        "500":
          description: 登录失败
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/v1/auth/send-sms:
    post:
      tags: [认证]
      summary: 发送短信验证码
      description: 向手机号发送 6 位数字验证码，有效期 5 分钟。开发模式下会在响应中返回 devCode。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [phoneNumber]
              properties:
                phoneNumber:
                  type: string
                  pattern: "^1\\d{10}$"
                  example: "13800138000"
                purpose:
                  type: string
                  enum: [login, bind_phone]
                  default: login
      responses:
        "200":
          description: 发送成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  expiresInSeconds:
                    type: integer
                    example: 300
                  devCode:
                    type: string
                    description: 仅开发模式返回
                    example: "123456"
  /api/v1/auth/sms-login:
    post:
      tags: [认证]
      summary: 手机号验证码登录
      description: 使用手机号 + 验证码登录，手机号不存在时自动注册。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [phoneNumber, code]
              properties:
                phoneNumber:
                  type: string
                  example: "13800138000"
                code:
                  type: string
                  example: "123456"
      responses:
        "200":
          description: 登录成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                  user:
                    $ref: "#/components/schemas/User"
                  children:
                    type: array
                    items:
                      $ref: "#/components/schemas/ChildProfile"
  /api/v1/auth/bind-phone:
    post:
      tags: [认证]
      summary: 绑定手机号
      description: 已登录用户绑定手机号，需先获取 purpose=bind_phone 的验证码。
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [phoneNumber, code]
              properties:
                phoneNumber:
                  type: string
                code:
                  type: string
      responses:
        "200":
          description: 绑定成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    $ref: "#/components/schemas/User"
                  children:
                    type: array
                    items:
                      $ref: "#/components/schemas/ChildProfile"
        "409":
          description: 手机号已被其他账号绑定
  /api/v1/auth/profile:
    get:
      tags: [认证]
      summary: 获取用户资料
      security:
        - bearerAuth: []
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    $ref: "#/components/schemas/User"
                  children:
                    type: array
                    items:
                      $ref: "#/components/schemas/ChildProfile"
    put:
      tags: [认证]
      summary: 更新用户资料
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                nickname:
                  type: string
                avatarUrl:
                  type: string
                gender:
                  type: string
                  enum: [male, female, undisclosed]
      responses:
        "200":
          description: 更新成功
  /api/v1/auth/child:
    post:
      tags: [认证]
      summary: 创建宝贝信息
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [nickname]
              properties:
                nickname:
                  type: string
                  example: "小星星"
                birthYear:
                  type: integer
                  example: 2020
                birthMonth:
                  type: integer
                  example: 6
                gender:
                  type: string
                  enum: [male, female, undisclosed]
      responses:
        "200":
          description: 创建成功
  /api/v1/auth/child/{childId}:
    put:
      tags: [认证]
      summary: 更新宝贝信息
      security:
        - bearerAuth: []
      parameters:
        - name: childId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                nickname:
                  type: string
                birthYear:
                  type: integer
                birthMonth:
                  type: integer
                gender:
                  type: string
                avatarUrl:
                  type: string
      responses:
        "200":
          description: 更新成功
  /api/v1/auth/logout:
    post:
      tags: [认证]
      summary: 退出登录
      security:
        - bearerAuth: []
      responses:
        "200":
          description: 成功

  # ========== 内容管理 ==========
...